Rearrange components into components folder and finish out project
