+++ src/components/IdeaCard/IdeaCard.js
@@ -1,12 +1,12 @@
 import React from 'react';
 
 const IdeaCard = (props) => {
-  const {title, body, id} = props.idea
+  const {title, body, id} = props.idea;
   return (
     <div>
       <h2>{title}</h2>
       <p>{body}</p>
-      <button onClick={() => props.removeIdea(id)}>X</button>
+      <button onClick={() => props.removeIdea(id)}>x</button>
     </div>
   )
 }
