+++ src/reducers/ideas-reducer.js
@@ -1,12 +1,12 @@
-const initialState = [];
+const initialState = []
 
 const ideasReducer = (state=initialState, action) => {
   switch(action.type) {
-    case 'CREATE_IDEA' :
+    case 'CREATE_IDEA':
       return [...state, action.newIdea]
-    case 'DELETE_IDEA' :
+    case 'DELETE_IDEA':
       return action.ideas
-    default :
+    default:
       return state
   }
 }
